requirements.txt
alembic==1.8.1
attrs==22.1.0
bcrypt==4.0.1
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.0.4
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Bcrypt==1.0.1
Flask-Ext==0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.0
flask-peewee==3.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
google-auth==2.15.0
google-auth-oauthlib==0.8.0
greenlet==2.0.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
oauthlib==3.2.2
packaging==22.0
peewee==3.15.4
pluggy==1.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.44
stripe==5.0.0
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
wtf-peewee==3.0.4
WTForms==3.0.1
WTForms-Appengine==0.1
